+++ src/Client/Views/ProtocolInsertView.fs
@@ -305,13 +305,12 @@ let addFromDBToTableButton (model:Model) dispatch =
                                 let version = p.Version
                                 let swateVersion = model.PersistentStorageState.AppVersion
                                 GroupTypes.Protocol.create id version swateVersion [] "" ""
-                            let minBuildingBlockInfos =
-                                model.ProtocolInsertState.BuildingBlockMinInfoList |> List.rev
+                            let minBuildingBlockInfos = model.ProtocolInsertState.BuildingBlockMinInfoList |> List.rev
                             /// Use x.Value |> Some to force an error if isNone. Otherwise AddAnnotationBlocks would just ignore it and it might be overlooked.
                             let validation =
                                 model.ProtocolInsertState.ValidationXml.Value |> Some
                             ProtocolIncreaseTimesUsed preProtocol.Id |> ProtocolInsert |> dispatch
-                            AddAnnotationBlocks (minBuildingBlockInfos,preProtocol, validation) |> ExcelInterop |> dispatch
+                            AddAnnotationBlocks (minBuildingBlockInfos, preProtocol, validation) |> ExcelInterop |> dispatch
                         )
                     ] [
                         str "Insert protocol annotation blocks"