+++ src/Client/Views/SettingsView.fs
@@ -60,6 +60,22 @@ let dataStewardsSettings (model:Model) dispatch =
         ]
     ]
 
+let protocolSettings (model:Model) dispatch =
+    Level.level [Level.Level.IsMobile][
+        Level.left [][
+            str Routing.Route.SettingsProtocol.toStringRdbl
+        ]
+        Level.right [ Props [ Style [if model.SiteStyleState.IsDarkMode then Color model.SiteStyleState.ColorMode.Text else Color model.SiteStyleState.ColorMode.Fade]]] [
+            Button.a [
+                Button.Color IsInfo
+                Button.IsOutlined
+                Button.OnClick (fun e -> UpdatePageState (Some Routing.Route.SettingsProtocol) |> dispatch ) 
+            ][
+                str "Advanced Settings"
+            ]
+        ]
+    ]
+
 let settingsViewComponent (model:Model) dispatch =
     div [
         //Style [MaxWidth "500px"]
@@ -73,4 +89,5 @@ let settingsViewComponent (model:Model) dispatch =
         Label.label [][str "Advanced Settings"]
         customXmlSettings model dispatch
         dataStewardsSettings model dispatch
+        protocolSettings model dispatch
     ]